refactor: 재시작 입력값을 담는 Restart 객체의 입력값 검증 기능을 Restart 내부로 이동
enum과 validate 메소드 하나만으로 입력값 검증이 가능하고, Restart 옵션 확장 시에도 enum 내부에서 처리 가능함.
따라서 굳이 RestartController를 두지 않고 Restart 내부로 이동.
