+++ src/main/java/baseball/domain/Restart.java
@@ -15,5 +15,13 @@ public enum Restart {
         return value;
     }
 
+    public static Restart validateValue(int value) {
+        for (Restart choice : Restart.values()) {
+            if (choice.getValue() == value) {
+                return choice;
+            }
+        }
+        return Restart.OTHER_CHOICE;
+    }
 }
